pyproject.toml
[tool.poetry]
name = "config-binder"
version = "0.2.1"
description = "Easily parse configurations across various formats (YAML, JSON, etc.) with recursive class binding, environment variables resolution and type safety."
authors = ["Pavel Istomin <inspaulation@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
